Fix warnings about comparison of unsigned and signed values.
git-svn-id: http://svn.boost.org/svn/boost/trunk@23892 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/mpl/test/lambda.cpp
@@ -44,7 +44,7 @@ int main()
                     mpl::not_< boost::is_same<_1, char> >
                   , mpl::not_< boost::is_float<_1> >
                   >
-            , mpl::greater< mpl::sizeof_<_1>, mpl::int_<8> >
+            , mpl::greater< mpl::sizeof_<_1>, mpl::size_t<8> >
             >
         >::type f1;
